NDLA-145 forenkling av migrering 21, behandler de gamle URLene uten å endre tabellstruktur
